+++ src/main/java/org/eclipse/tractusx/selfdescriptionfactory/service/SDFactory.java
@@ -25,6 +25,7 @@ import com.danubetech.verifiablecredentials.VerifiableCredential;
 import foundation.identity.jsonld.JsonLDUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.eclipse.tractusx.selfdescriptionfactory.service.clearinghouse.ClearingHouse;
 import org.eclipse.tractusx.selfdescriptionfactory.service.wallet.CustodianWallet;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.convert.ConversionService;
@@ -51,14 +52,16 @@ public class SDFactory {
     private int duration;
     private final CustodianWallet custodianWallet;
     private final ConversionService conversionService;
+    private final ClearingHouse clearingHouse;
 
     @PreAuthorize("hasRole(@securityRoles.createRole)")
-    public ResponseEntity<Map<String, Object>> createVC(Object document) {
+    public void createVC(Object document) {
         var claimsHolder = Optional.ofNullable(conversionService.convert(document, Claims.class)).orElseThrow();
         var claims = new HashMap<>(claimsHolder.claims());
         var holder = claims.remove("holder");
         var issuer = claims.remove("issuer");
         var type = claims.get("type");
+        var externalId = claims.remove("externalId");
         var credentialSubject = CredentialSubject.fromJsonObject(claims);
         var verifiableCredential = VerifiableCredential.builder()
                 .context(claimsHolder.vocabulary())
@@ -70,6 +73,6 @@ public class SDFactory {
         JsonLDUtils.jsonLdAdd(verifiableCredential, "holderIdentifier", holder);
         JsonLDUtils.jsonLdAdd(verifiableCredential, "type", type);
         var result = custodianWallet.getSignedVC(verifiableCredential);
-        return new ResponseEntity<>(result.toMap(), HttpStatus.CREATED);
+        clearingHouse.sendToClearingHouse(result, externalId.toString());
     }
 }