+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Changed
+- add async support. Remove 22.10 API from endpoint
 - schema was changed so registrationNumber is an array of documents
   with type and value fields  
 - removed support for 22.10 as OpenAPI yaml is buggy and there ara